+++ LaborVolumeCalculator/Models/Dictionary/OkrInnovationRate.cs
@@ -10,7 +10,7 @@ namespace LaborVolumeCalculator.Models.Dictionary
     {
         public OkrInnovationRate() { }
 
-        public OkrInnovationRate(OkrInnovationProperty okrInnovationProperty, DeviceComposition deviceComposition, decimal value): this()
+        public OkrInnovationRate(OkrInnovationProperty okrInnovationProperty, DeviceComposition deviceComposition, double value): this()
         {
             OkrInnovationProperty = okrInnovationProperty;
             OkrInnovationPropertyID = okrInnovationProperty.ID;
@@ -35,6 +35,6 @@ namespace LaborVolumeCalculator.Models.Dictionary
         public DeviceComposition DeviceComposition { get; set; }
 
         [DisplayName("Значение")]
-        public decimal Value { get; set; }
+        public double Value { get; set; }
     }
 }
